% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.test.R
\name{bf.test}
\alias{bf.test}
\title{Check for Brown-Forsythe test}
\usage{
bf.test(x, y)
}
\arguments{
\item{x}{is a set of vector.}

\item{y}{is another set of vector.}
}
\value{
Brown-Forsythe t-statistic and its p-value.
}
\description{
bf.test will give Brown-Forsythe t-statistic and its p-value.
}
\examples{
x=rnorm(100); y=rnorm(50); # Two set of vectors are generated from same normal distribution with mean=0, sd=1
bf.test(x,y)
}
