% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.trsf.R
\name{cor.trsf}
\alias{cor.trsf}
\title{Method for correlation transformation}
\usage{
cor.trsf(x)
}
\arguments{
\item{x}{is given vector}
}
\value{
cor. trsfed vector
}
\description{
cor.trsf function will do correlation transformation for given vector
}
\examples{
x=c(1,2,3)
cor.trsf(x)
}
