% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.sel.R
\name{mod.sel}
\alias{mod.sel}
\title{Provide a data.frame of useful statistics for some models}
\usage{
mod.sel(x, y, z)
}
\arguments{
\item{x}{is a set of independent variables.}

\item{y}{is a response variable}

\item{z}{is for nbest in regsubsets. Try several numbers, from 1 to choose(ncol(x),1).}
}
\value{
It provides p(number of parameters), SSE, R.squared, R.adj.squared, AIC, BIC and PRESS
}
\description{
This data.frame can be found in Applied Linear Regression Models, Kutner, Chapter 9. Criteria for model selection
}
\examples{
library(ALSM)
mod.sel(SurgicalUnit[,1:4],SurgicalUnit[,10],1)
mod.sel(SurgicalUnit[,1:6],SurgicalUnit[,10],2)
mod.sel(BodyFat[,1:3],BodyFat[,4],3)
}
