% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.valid.R
\name{mod.valid}
\alias{mod.valid}
\title{Model validation at a glance}
\usage{
mod.valid(a, b, c, d, e)
}
\arguments{
\item{a}{is a entire set of indep variables in model-building set.}

\item{b}{is a response variable in model-building set.}

\item{c}{is a entire set of indep variables in validation set.}

\item{d}{is a response variable in validation set}

\item{e}{is a number of column(s) you want to put in the model. ex) lm(y~x1+x2+x3+x8) is what you want, e=c(1,2,3,8)}
}
\value{
a matrix with coefficients, p, Mallow's Cp, R.squared, R.adj.squared, SSE, PRESS, MSE, MSPR, AIC and BIC
}
\description{
mod.valid function will provide internal&external statistics for model validation.
}
\examples{
library(ALSM)
mod.valid(SurgicalUnit[,1:8],SurgicalUnit[,10],SurgicalUnitAdditional[,1:8],SurgicalUnitAdditional[,10],c(1,2,3,8))
}
